move storage validator in separate package
