+++ tests/test_device_validator.py
@@ -22,13 +22,13 @@ import pytest
 
 from d3a_interface.constants_limits import ConstSettings
 from d3a_interface.device_validator import (
-    validate_storage_device,
     validate_commercial_producer, validate_market_maker, validate_finite_diesel_generator,
     validate_infinite_bus)
 from d3a_interface.exceptions import D3ADeviceException
 from d3a_interface.validators.home_meter_validator import HomeMeterValidator
 from d3a_interface.validators.load_validator import validate_load_device
 from d3a_interface.validators.pv_validator import validate_pv_device
+from d3a_interface.validators.storage_validator import validate_storage_device
 
 GeneralSettings = ConstSettings.GeneralSettings
 LoadSettings = ConstSettings.LoadSettings
